* fishmanpet/umnmaster:
  Update BuildHelpers.Tests.ps1
  Update Start-Build.ps1
  Update BuildHelpers.Tests.ps1
  update bh
  Update Start-Build.ps1
  Update build.requirements.psd1
  Update psake.ps1
  Update BuildHelpers.Tests.ps1
  Update BuildHelpers.Tests.ps1
  Update BuildHelpers.Tests.ps1
  I can count
  Update Get-ProjectName.ps1
  Update Get-ProjectName.ps1
  better project name detection when CI doesn't use sensible name as root dir
  Update Get-BuildVariable.ps1
  Update Get-BuildVariable.ps1
  Raw Revision String
  Filtering for days
  Update Get-GitChangedFile.ps1

+++ Build/psake.ps1
@@ -45,7 +45,12 @@ Task Test -Depends Init  {
             "$ProjectRoot\$TestFile" )
     }
 
-    Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+    #Azure Pipelines requires the test files to stay around so they can be uploaded
+    #If you're running this locally ('Unknown') you might want the results to stick around
+    If($ENV:BHBuildSystem -notin 'Azure Pipelines','Unknown')
+    {
+        Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+    }
 
     # Failed tests?
     # Need to tell psake or it will proceed to the deployment. Danger!
